Make `castOrFail` function public to support class cluster for now
